Fixes #61, just a refactoring of the http client
