+++ Core/Exercises/Contracts/Repositories/IExerciseRepository.cs
@@ -1,0 +1,9 @@
+﻿using Core.Exercises.Models;
+using FluentResults;
+
+namespace Core.Exercises.Contracts.Repositories;
+
+public interface IExerciseRepository
+{
+    Task<Result> InsertExerciseAsync(ExerciseDto dto, int userId);
+}
