+++ Core/Exercises/Contracts/Services/ISolutionRunnerService.cs
@@ -1,0 +1,9 @@
+﻿using Core.Exercises.Models;
+
+namespace Core.Exercises.Contracts.Services;
+
+public interface ISolutionRunnerService
+{
+    Task SubmitSolutionAsync(ExerciseDto dto);
+
+}
